add apptext and ui changes
